+++ mtgcc-spoiler-bot/modules/generic/return_channels.py
