+++ mtgcc-spoiler-bot/modules/generic/time_functions.py
