+++ mtgcc-spoiler-bot/modules/spoiler_modules/html_processing.py
