+++ mtgcc-spoiler-bot/modules/spoiler_modules/open_mythicspoier.py
