+++ mtgcc-spoiler-bot/modules/spoiler_modules/read_set_file.py
