+++ mtgcc-spoiler-bot/modules/spoiler_modules/set_file_creation_funcions.py
