+++ mtgcc-spoiler-bot/set_data/tdm.json
@@ -589,3 +589,14 @@ piercingexhale: tdm/cards/piercingexhale.jpg
 krotiqnestguard: tdm/cards/krotiqnestguard.jpg
 roarofendlesssong: tdm/cards/roarofendlesssong.jpg
 hardenedtactician: tdm/cards/hardenedtactician.jpg
+wayspeakerbodyguard: tdm/cards/wayspeakerbodyguard.jpg
+amblingstormshell: tdm/cards/amblingstormshell.jpg
+highspirebellringer: tdm/cards/highspirebellringer.jpg
+deltabloodflies: tdm/cards/deltabloodflies.jpg
+starofthearena: tdm/cards/starofthearena.jpg
+zurgosvanguard: tdm/cards/zurgosvanguard.jpg
+shockingsharpshooter: tdm/cards/shockingsharpshooter.jpg
+travelingbotanist: tdm/cards/travelingbotanist.jpg
+mammothbellow: tdm/cards/mammothbellow.jpg
+lotuslightdancers: tdm/cards/lotuslightdancers.jpg
+defibrillatingcurrent: tdm/cards/defibrillatingcurrent.jpg
